feat(vue-smart-pages): support watch with throttle
