+++ packages/vue-smart-pages/src/index.ts
@@ -1,4 +1,6 @@
 import { useSmartPages } from "./useSmartPages";
 
-export default useSmartPages;
 export * from "./useSmartPages";
+export type * from "./types";
+
+export default useSmartPages;
