+++ packages/vue-smart-pages/src/types.ts
@@ -1,0 +1,11 @@
+export type PageSize = {
+  width: number;
+  height: number;
+};
+
+export type PageMargins = Partial<{
+  top: number;
+  bottom: number;
+  left: number;
+  right: number;
+}>;
