+++ site/src/components/shared/ResumeRender.vue
@@ -32,6 +32,9 @@ const { render } = useSmartPages(target, html, size, margins, {
   beforeRender: async () => {
     // Wait for the fonts to be loaded
     await googleFontsService.presetObserver(props.styles);
+  },
+  watchThrottledOptions: {
+    throttle: 200
   }
 });
